Changed some file names. Added some comments and documentation.
