+++ tst/run_test_suite.py
@@ -1,0 +1,79 @@
+#!/usr/bin/env python
+
+"""
+Script to run automatic test suite.
+
+Usage: From this directory, call this script with python:
+      python run_test_suite.py [ARGS]
+
+Notes:
+  - Arguments specify whether to run tests on CPU, CPU+MPI, or GPU
+  - May need to pass additional arguments for CMake (especially on GPUs)
+  - Requires Python 2.7+. (compliant with Python 3)
+  - This file does not need to be modified when adding new scripts.
+  - To add a new script, create a new .py file in a /test_suite/ subdirectory.
+  - Scripts that run tests on CPU must have '_cpu' in name
+  - Scripts that run tests on CPU with MPI must have '_mpicpu' in name
+  - Scripts that run tests on GPU must have '_gpu' in name
+  - For more information, check online automatic testing wiki page.
+"""
+
+import sys
+sys.path.append('../tst/test_suite')
+import test_suite.testutils as testutils
+import pytest
+import argparse
+
+def cmake_flags(args,flags):
+    """Process command line arguments and return cmake flags."""
+    if args:
+        for arg in args:
+            flags += (arg.split(" "))
+    return flags
+
+# Set up argument parser
+parser = argparse.ArgumentParser(description="Run AthenaK test suite.")
+parser.add_argument(
+    "--style", 
+    help="check source code conforms to style guide."
+)
+parser.add_argument(
+    "--cpu", 
+    nargs="*", 
+    help="Run test on CPU. Can add optional cmake arguments."
+)
+parser.add_argument(
+    "--mpicpu", 
+    nargs="*", 
+    help="Run test on CPU with MPI. Can add additional cmake arguments."
+)
+parser.add_argument(
+    "--gpu", 
+    nargs="*", 
+    help="Run test on GPU. Can add optional cmake arguments."
+)
+
+args = parser.parse_args()
+status=True
+for arg in vars(args):
+    status *= getattr(args,arg) == None
+if status:
+    print("No target device (CPU/GPU) specified.")
+    print(parser.format_help())
+    sys.exit(1)
+
+# Run tests based on arguments
+if  args.style:
+    pytest.main(["test_suite/style"])
+
+if args.cpu != None:
+    testutils.clean_make(flags=cmake_flags(args.cpu, []))
+    pytest.main(["test_suite", "-k", "_cpu"])
+
+if args.mpicpu != None: 
+    testutils.clean_make(flags=cmake_flags(args.mpicpu, ["-D", "Athena_ENABLE_MPI=ON"]))
+    pytest.main(["test_suite", "-k", "_mpicpu"])
+if args.gpu != None: 
+    testutils.clean_make(flags=cmake_flags(args.gpu, ["-D", "Kokkos_ENABLE_CUDA=On"]),text=True)
+    pytest.main(["test_suite", "-k", "_gpu"])
+testutils.clean()
