+++ tst/test_suite/gr/test_gr_lwave3d_amr_gpu.py
@@ -8,9 +8,9 @@
 
 import sys
 sys.path.insert(0, '../vis/python')
-sys.path.insert(0, '../tests_suite')
+sys.path.insert(0, '../test_suite')
 import pytest
-import tests_suite.testutils as testutils
+import test_suite.testutils as testutils
 import scripts.utils.athena as athena
 import athena_read
 import numpy as np
@@ -18,34 +18,33 @@ import numpy as np
 # Threshold errors and convergence rates
 # for different integrators, reconstructions, and wave types
 maxerrors={
-    ('hydro', 'rk3', 'plm', '0'): (3.7e-09,0.25),
-    ('hydro', 'rk3', 'ppm4', '0'): (2e-09,0.25),
-    ('hydro', 'rk3', 'ppmx', '0'): (1.8e-09,0.39),
-    ('hydro', 'rk3', 'wenoz', '0'): (3.3e-10,0.24),
-    ('mhd', 'rk3', 'plm', '0'): (1.1e-08,0.24),
-    ('mhd', 'rk3', 'ppm4', '0'): (5e-09,0.22),
-    ('mhd', 'rk3', 'ppmx', '0'): (3.7e-09,0.3),
-    ('mhd', 'rk3', 'wenoz', '0'): (9.1e-10,0.24),}
+    ('hydro', 'rk3', 'plm', '0'): (1.5e-08,0.26),
+    ('hydro', 'rk3', 'ppm4', '0'): (7.8e-09,0.24),
+    ('hydro', 'rk3', 'ppmx', '0'): (4.7e-09,0.24),
+    ('hydro', 'rk3', 'wenoz', '0'): (1.5e-09,0.23),
+    ('mhd', 'rk3', 'plm', '0'): (4.6e-08,0.26),
+    ('mhd', 'rk3', 'ppm4', '0'): (2.3e-08,0.23),
+    ('mhd', 'rk3', 'ppmx', '0'): (1.3e-08,0.26),
+    ('mhd', 'rk3', 'wenoz', '0'): (3.9e-09,0.23),}
 
 _int = ['rk3']
 _recon = ['wenoz']  # do not change order
 _wave = ['0']                      # do not change order
 _flux = ['hlle']
-_res = [32, 64]                            # resolutions to test
+_res = [16, 32]                            # resolutions to test
 _soe = ['hydro', 'mhd']  # system of equations to test
 
-def arguments(iv, rv, fv, wv, res,soe,name):
-    vflow = 1.0 if wv=='3' else 0.0
+def arguments(iv, rv, fv, wv, res, soe, name, dim):
     return  [f'job/basename={name}',
             'time/tlim=1.0',
             'time/integrator=' + iv,
             'mesh/nghost=4',
             'mesh/nx1=' + repr(res),
-            'mesh/nx2=' + repr(res//2),
-            'mesh/nx3=1',
-            'meshblock/nx1=16',
-            'meshblock/nx2=16',
-            'meshblock/nx3=1',
+            'mesh/nx2=' + repr(res//2 if dim > 1 else 1),
+            'mesh/nx3=' + repr(res//4 if dim > 2 else 1),
+            'meshblock/nx1=4',
+            'meshblock/nx2='+ repr(4 if dim > 1 else 1),
+            'meshblock/nx3='+ repr(4 if dim > 2 else 1),
             'time/cfl_number=0.4',
             'coord/special_rel=false',
             'coord/general_rel=true',
@@ -55,20 +54,20 @@ def arguments(iv, rv, fv, wv, res,soe,name):
 
 
 @pytest.mark.parametrize("iv" , _int)
+@pytest.mark.parametrize("dim" , [1,2,3])
 @pytest.mark.parametrize("rv" , _recon)
 @pytest.mark.parametrize("fv" , _flux)
 @pytest.mark.parametrize("soe" , _soe)
-def test_run(iv, fv, rv, soe):
+def test_run(iv, fv, rv, soe, dim):
     """Run a single test with given parameters."""
-    l1_rms_l,l1_rms_r = testutils.test_error_convergence(
-            f"inputs/lwave_rel{soe}.athinput",
-            f"sr_lwave_{soe}",
-            arguments,
-            maxerrors,
-            _wave,
-            _res,
-            iv,
-            rv,
-            fv,
-            soe,
-            )
+    _,_ = testutils.test_error_convergence(
+        f"inputs/lwave_rel{soe}.athinput",
+        f'SR_{soe}_linwave{dim}d_amr',
+        lambda iv, rv, fv, wv, res, soe, name : arguments(iv, rv, fv, wv, res, soe, name, dim=dim),
+        maxerrors,
+        _wave,
+        _res,
+        iv,
+        rv,
+        fv,
+        soe,)
