+++ tst/test_suite/nr/test_lwave2d_amr_cpu.py
@@ -8,9 +8,9 @@
 
 import sys
 sys.path.insert(0, '../vis/python')
-sys.path.insert(0, '../tests_suite')
+sys.path.insert(0, '../test_suite')
 import pytest
-import tests_suite.testutils as testutils
+import test_suite.testutils as testutils
 import scripts.utils.athena as athena
 import athena_read
 import numpy as np
