+++ tst/test_suite/sr/test_sr_shocktube_cpu.py
@@ -8,9 +8,9 @@
 import sys
 import os
 sys.path.append('../vis/python')
-sys.path.append('../tst/tests_suite')
+sys.path.append('../tst/test_suite')
 import pytest
-import tests_suite.testutils as testutils
+import test_suite.testutils as testutils
 import scripts.utils.athena as athena
 import athena_read
 import numpy as np
