+++ tst/test_suite/style/check_athena_cpp_style.sh
