+++ tst/test_suite/style/clang_format_v1.cfg
