+++ tst/test_suite/style/clang_format_v2.cfg
